[Feature] NumberFormatter manager
